requirements.txt

Flask==2.3.2
Flask-MySQLdb==0.2.0
python-dotenv==1.0.0
